fixed PSO problem

+++ src/Particule.java
@@ -1,10 +1,13 @@
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 public class Particule {
 
     int[] position;
-    float velocity;
+    BigInteger velocity;
     int score;
 
-    public Particule(int[] position, float velocity) {
+    public Particule(int[] position, BigInteger velocity) {
         this.position = position;
         this.velocity = velocity;
     }
